services/commentary/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.74.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
earthengine-api==1.4.2
